Move to INamedTypeSymbolExtensionsSonar
